+++ container_ci_suite/openshift.py
@@ -28,10 +28,12 @@ import random
 import subprocess
 
 from pathlib import Path
-from subprocess import CalledProcessError
 from typing import Dict, Any, List
 
 from container_ci_suite.container import DockerCLIWrapper
+from container_ci_suite.openshift_ops import OpenShiftOperations
+from container_ci_suite.utils import run_oc_command
+
 
 import container_ci_suite.utils as utils
 
@@ -47,62 +49,45 @@ class OpenShiftAPI:
             delete_prj: bool = True,
             version: str = ""
     ):
-        self.namespace = namespace
         self.create_prj = create_prj
         self.delete_prj = delete_prj
         self.shared_cluster = utils.is_share_cluster()
         self.pod_name_prefix = pod_name_prefix
         self.pod_json_data: Dict = {}
         self.version = version
-        self.build_failed: bool = False
         self.shared_random_name = ""
         self.config_tenant_name = "core-services-ocp--config"
-        if self.namespace == "default":
+        self.openshift_ops = OpenShiftOperations(pod_name_prefix=pod_name_prefix)
+        if namespace == "default":
             self.create_project()
         else:
             self.namespace = namespace
             self.create_prj = False
-
-    @staticmethod
-    def run_oc_command(
-        cmd, json_output: bool = True, return_output: bool = True, ignore_error: bool = False, shell: bool = True,
-            namespace: str = ""
-    ):
-        """
-        Run docker command:
-        """
-        json_cmd = "-o json" if json_output else ""
-        namespace_cmd = f"-n {namespace}" if namespace != "" else ""
-
-        return utils.run_command(
-            f"oc {cmd} {namespace_cmd} {json_cmd}",
-            return_output=return_output,
-            ignore_error=ignore_error,
-            shell=shell,
-        )
 
     def create_project(self):
         if self.create_prj:
             if self.shared_cluster:
                 self.shared_random_name = f"{random.randrange(10000, 100000)}"
                 self.namespace = f"core-services-ocp--{self.shared_random_name}"
+                self.openshift_ops.set_namespace(self.namespace)
                 self.prepare_tenant_namespace()
             else:
                 self.namespace = f"sclorg-{random.randrange(10000, 100000)}"
-                OpenShiftAPI.run_oc_command(f"new-project {self.namespace}", json_output=False, return_output=True)
+                self.openshift_ops.set_namespace(self.namespace)
+                run_oc_command(f"new-project {self.namespace}", json_output=False, return_output=True)
         else:
-            OpenShiftAPI.run_oc_command(f"project {self.namespace}", json_output=False)
-        return self.is_project_exits()
+            run_oc_command(f"project {self.namespace}", json_output=False)
+        return self.openshift_ops.is_project_exits()
 
     def prepare_tenant_namespace(self):
         json_flag = False
         self.login_to_shared_cluster()
         tenant_yaml_file = utils.save_tenant_namespace_yaml(project_name=self.shared_random_name)
-        OpenShiftAPI.run_oc_command(cmd=f"create -f {tenant_yaml_file}", json_output=json_flag, return_output=True)
+        run_oc_command(cmd=f"create -f {tenant_yaml_file}", json_output=json_flag, return_output=True)
         tenant_egress_file = utils.save_tenant_egress_yaml(project_name=self.shared_random_name)
-        OpenShiftAPI.run_oc_command(cmd=f"apply -f {tenant_egress_file}", json_output=False, return_output=True)
+        run_oc_command(cmd=f"apply -f {tenant_egress_file}", json_output=False, return_output=True)
         time.sleep(3)
-        OpenShiftAPI.run_oc_command(
+        run_oc_command(
             cmd=f"project {self.namespace}",
             json_output=json_flag,
             return_output=True
@@ -110,12 +95,12 @@ class OpenShiftAPI:
 
     def delete_tenant_namespace(self):
         json_flag = False
-        namespace = OpenShiftAPI.run_oc_command(cmd="project -q", json_output=json_flag)
+        namespace = run_oc_command(cmd="project -q", json_output=json_flag)
         if namespace == self.config_tenant_name:
             print(f"Deleting tenant '{self.config_tenant_name}' is not allowed.")
             return
-        OpenShiftAPI.run_oc_command(f"project {self.config_tenant_name}", json_output=json_flag)
-        if OpenShiftAPI.run_oc_command(
+        run_oc_command(f"project {self.config_tenant_name}", json_output=json_flag)
+        if run_oc_command(
                 f"delete tenantnamespace {self.shared_random_name}",
                 json_output=json_flag
         ):
@@ -131,231 +116,35 @@ class OpenShiftAPI:
         if self.shared_cluster:
             self.delete_tenant_namespace()
         else:
-            OpenShiftAPI.run_oc_command("project default", json_output=False)
-            OpenShiftAPI.run_oc_command(
+            run_oc_command("project default", json_output=False)
+            run_oc_command(
                 f"delete project {self.namespace} --grace-period=0 --force", json_output=False
             )
 
-    def is_project_exits(self) -> bool:
-        output = OpenShiftAPI.run_oc_command("projects", json_output=False)
-        if self.namespace in output:
-            return True
-        return False
-
-    def get_pod_count(self) -> int:
-        count: int = 0
-        for item in self.pod_json_data["items"]:
-            pod_name = item["metadata"]["name"]
-            if self.pod_name_prefix not in pod_name:
-                continue
-            if "deploy" in pod_name:
-                continue
-            if "build" in pod_name:
-                continue
-            count += 1
-        return count
-
-    def is_pod_running(self, pod_name_prefix: str = "", loops: int = 180) -> bool:
-        print(f"Check for POD is running {pod_name_prefix}")
-        for count in range(loops):
-            print(".", sep="", end="")
-            self.pod_json_data = self.get_pod_status()
-            if pod_name_prefix == "" and self.pod_name_prefix == "":
-                print("\nApplication pod name is not specified. Call: is_pod_running(pod_name_prefix=\"something\").")
-                return False
-            if pod_name_prefix != "":
-                self.pod_name_prefix = pod_name_prefix
-            # Only one running pod is allowed
-            if self.get_pod_count() != 1:
-                time.sleep(3)
-                continue
-
-            for item in self.pod_json_data["items"]:
-                pod_name = item["metadata"]["name"]
-                if self.pod_name_prefix not in pod_name:
-                    continue
-                status = item["status"]["phase"]
-                if "deploy" in pod_name:
-                    print(".", sep="", end="")
-                    continue
-                if item["status"]["phase"] == "Running":
-                    print(f"\nPod with name {pod_name} is running {status}.")
-                    output = OpenShiftAPI.run_oc_command(
-                        f"logs {pod_name}", namespace=self.namespace, json_output=False
-                    )
-                    print(output)
-                    # Wait couple seconds for sure
-                    time.sleep(3)
-                    return True
-            time.sleep(3)
-        return False
-
-    def is_build_pod_finished(self, cycle_count: int = 180) -> bool:
-        """
-        Function return information if build pod is finished.
-        The function waits for 180*3 seconds
-        """
-        print("Check if build pod is finished")
-        for count in range(cycle_count):
-            print(".", sep="", end="")
-            self.pod_json_data = self.get_pod_status()
-            if len(self.pod_json_data["items"]) == 0:
-                time.sleep(3)
-                continue
-            if not self.is_build_pod_present():
-                print(".", sep="", end="")
-                time.sleep(3)
-                continue
-            if not self.is_pod_finished(pod_suffix_name="build"):
-                print(".", sep="", end="")
-                if self.build_failed:
-                    return False
-                time.sleep(3)
-                continue
-            print("\nBuild pod is finished")
-            return True
-        return False
-
-    def is_s2i_pod_running(self, pod_name_prefix: str = "", cycle_count: int = 180) -> bool:
-        self.pod_name_prefix = pod_name_prefix
-        build_pod_finished = False
-        print("Check if S2I build pod is running")
-        for count in range(cycle_count):
-            print(".", sep="", end="")
-            self.pod_json_data = self.get_pod_status()
-            if len(self.pod_json_data["items"]) == 0:
-                time.sleep(3)
-                continue
-            if not self.is_build_pod_present():
-                time.sleep(3)
-                continue
-            if not self.is_pod_finished(pod_suffix_name="build"):
-                time.sleep(3)
-                continue
-            build_pod_finished = True
-            print(f"\nBuild pod with name {pod_name_prefix} is finished.")
-        if not build_pod_finished:
-            print(f"\nBuild pod with name {pod_name_prefix} was not finished.")
-            return False
-        print("Check if S2I pod is running.")
-        for count in range(cycle_count):
-            print(".", sep="", end="")
-            if not self.is_pod_running():
-                time.sleep(3)
-                continue
-            print("\nPod is running")
-            return True
-        return False
-
-    @staticmethod
-    def get_raw_url_for_json(container: str, dir: str, filename: str, branch: str = "master") -> str:
-        RAW_SCL_JSON_URL: str = "https://raw.githubusercontent.com/sclorg/{container}/{branch}/{dir}/{filename}"
-        return RAW_SCL_JSON_URL.format(container=container, branch=branch, dir=dir, filename=filename)
-
-    def get_pod_status(self) -> Dict:
-        # output = OpenShiftAPI.run_oc_command("get all", json_output=False)
-        # print(f"oc get all: {output}")
-        output = OpenShiftAPI.run_oc_command("get pods", json_output=True, namespace=self.namespace)
-        # print(f" oc get pods: {output}")
-        return json.loads(output)
-
-    def is_build_pod_present(self) -> bool:
-        for item in self.pod_json_data["items"]:
-            pod_name = item["metadata"]["name"]
-            if "build" in pod_name:
-                return True
-        return False
-
-    def print_get_status(self):
-        print("Print get all and status:")
-        print(OpenShiftAPI.run_oc_command("get all", namespace=self.namespace, json_output=False))
-        print(OpenShiftAPI.run_oc_command("status", namespace=self.namespace, json_output=False))
-        print(OpenShiftAPI.run_oc_command("status --suggest", namespace=self.namespace, json_output=False))
-
-    def print_pod_logs(self):
-        self.pod_json_data = self.get_pod_status()
-        print("Print all pod logs")
-        for item in self.pod_json_data["items"]:
-            pod_name = item["metadata"]["name"]
-            print(f"Logs from pod name {pod_name}:")
-            oc_logs = OpenShiftAPI.run_oc_command(f"logs pod/{pod_name}", json_output=False)
-            print(oc_logs)
-
-    def is_pod_finished(self, pod_suffix_name: str = "deploy") -> bool:
-        if not self.pod_json_data:
-            self.pod_json_data = self.get_pod_status()
-        for item in self.pod_json_data["items"]:
-            print(".", sep="", end="")
-            pod_name = item["metadata"]["name"]
-            if self.pod_name_prefix not in pod_name:
-                print(".", sep="", end="")
-                continue
-            status = item["status"]["phase"]
-            if pod_suffix_name in pod_name and status == "Failed":
-                print(f"\nPod with {pod_suffix_name} finished with {status}. See logs.")
-                self.build_failed = True
-                self.print_pod_logs()
-                self.print_get_status()
-                return False
-            if pod_suffix_name in pod_name and status != "Succeeded":
-                print(".", sep="", end="")
-                continue
-            if pod_suffix_name in pod_name and status == "Succeeded":
-                print(f"\nPod with suffix {pod_suffix_name} is finished")
-                return True
-        return False
-
     def run_command_in_pod(self, pod_name, command: str = "") -> str:
-        output = OpenShiftAPI.run_oc_command(f"exec {pod_name} -- \"{command}\"")
+        output = run_oc_command(f"exec {pod_name} -- \"{command}\"")
         print(output)
         return output
 
-    def oc_get_services(self, service_name):
-        output = OpenShiftAPI.run_oc_command(f"get svc/{service_name}", json_output=True, namespace=self.namespace)
-        json_output = json.loads(output)
-        print(json_output)
-        return json_output
-
-    def get_service_ip(self, service_name) -> Any:
-        json_output = self.oc_get_services(service_name=service_name)
-        if "clusterIP" not in json_output["spec"]:
-            return None
-        if not json_output["spec"]["clusterIP"]:
-            return None
-        return json_output["spec"]["clusterIP"]
-
-    def is_imagestream_exist(self, name: str):
-        try:
-            json_output = self.oc_get_is(name=name)
-            if json_output["kind"] == "ImageStream" and json_output["metadata"]["name"] == name:
-                return json_output
-        except CalledProcessError:
-            pass
-        return None
-
     def import_is(self, path: str, name: str, skip_check=False):
         if not skip_check:
-            is_exists = self.is_imagestream_exist(name=name)
+            is_exists = self.openshift_ops.is_imagestream_exist(name=name)
             if is_exists:
                 return is_exists
-        output = OpenShiftAPI.run_oc_command(f"create -f {path}", namespace=self.namespace)
+        output = run_oc_command(f"create -f {path}", namespace=self.namespace)
         # Let's wait 3 seconds till imagestreams are not uploaded
         time.sleep(3)
         return json.loads(output)
 
     def process_file(self, path: str):
-        output = OpenShiftAPI.run_oc_command(f"process -f {path}", namespace=self.namespace)
+        output = run_oc_command(f"process -f {path}", namespace=self.namespace)
         json_output = json.loads(output)
         print(json_output)
         return json_output
 
-    def oc_get_is(self, name: str):
-        output = OpenShiftAPI.run_oc_command(f"get is/{name}", namespace=self.namespace)
-        return json.loads(output)
-
     def start_build(self, service_name: str, app_name: str = "") -> str:
         from_dir = utils.download_template(template_name=app_name)
-        output = OpenShiftAPI.run_oc_command(f"start-build {service_name} --from-dir={from_dir}", json_output=False)
+        output = run_oc_command(f"start-build {service_name} --from-dir={from_dir}", json_output=False)
         return output
 
     def login_to_shared_cluster(self):
@@ -364,11 +153,11 @@ class OpenShiftAPI:
         if not all([token, url]):
             print("Important variables 'SHARED_CLUSTER_TOKEN,SHARED_CLUSTER_URL' are missing.")
             return None
-        output = OpenShiftAPI.run_oc_command(f"login --token={token} --server={url}", json_output=False)
+        output = run_oc_command(f"login --token={token} --server={url}", json_output=False)
         print(output)
-        output = OpenShiftAPI.run_oc_command("version", json_output=False)
+        output = run_oc_command("version", json_output=False)
         print(output)
-        output = OpenShiftAPI.run_oc_command(f"project {self.config_tenant_name}", json_output=False)
+        output = run_oc_command(f"project {self.config_tenant_name}", json_output=False)
         print(output)
 
     @staticmethod
@@ -401,7 +190,7 @@ class OpenShiftAPI:
         cmd = f"podman push {output_name}"
         output = utils.run_command(cmd, ignore_error=False, return_output=True)
         print(f"Output from podman push command {output}")
-        ret = OpenShiftAPI.run_oc_command(
+        ret = run_oc_command(
             f"import-image {tagged_image} --from={output_name} --confirm", json_output=False, return_output=True
         )
         print(ret)
@@ -410,14 +199,14 @@ class OpenShiftAPI:
         return True
 
     def docker_login_to_openshift(self) -> Any:
-        output = OpenShiftAPI.run_oc_command(cmd="get route default-route -n openshift-image-registry")
+        output = run_oc_command(cmd="get route default-route -n openshift-image-registry")
         jsou_output = json.loads(output)
         print(jsou_output["spec"]["host"])
         if not jsou_output["spec"]["host"]:
             print("Default route does not exist. Install OpenShift 4 cluster properly and expose default route.")
             return None
         ocp4_register = jsou_output["spec"]["host"]
-        token_output = OpenShiftAPI.run_oc_command(cmd="whoami -t", json_output=False).strip()
+        token_output = run_oc_command(cmd="whoami -t", json_output=False).strip()
         cmd = f"docker login -u kubeadmin -p {token_output} {ocp4_register}"
         output = utils.run_command(
             cmd=cmd,
@@ -459,25 +248,6 @@ class OpenShiftAPI:
         print(f"Upload_image push {output}")
         return True
 
-    def check_is_exists(self, is_name, version_to_check: str) -> bool:
-        """
-        Function checks if it exists in OpenShift 4 environment
-        Exact version has to be the same
-        :return:    True if tag was found
-                    False if tag does not exist
-        """
-        json_output = self.oc_get_is(name=is_name)
-        if "tags" not in json_output["spec"]:
-            return False
-        tag_found: bool = False
-        for tag in json_output["spec"]["tags"]:
-            if "name" not in tag:
-                continue
-            if version_to_check not in tag["name"]:
-                continue
-            tag_found = True
-        return tag_found
-
     def create_new_app_with_template(self, name: str, template_json: str, template_args: Dict = None):
         """
         Function creates a new application in OpenShift 4 environment
@@ -492,7 +262,7 @@ class OpenShiftAPI:
         args = [""]
         if template_args:
             args = [f"-p {key}={val}" for key, val in template_args]
-        output = OpenShiftAPI.run_oc_command(
+        output = run_oc_command(
             f"new-app {template_json} --name {name} -p NAMESPACE={self.namespace} {args}",
             json_output=False
         )
@@ -500,7 +270,7 @@ class OpenShiftAPI:
         return output
 
     def get_route_url(self, routes_name: str) -> Any:
-        output = OpenShiftAPI.run_oc_command(f"get routes/{routes_name}")
+        output = run_oc_command(f"get routes/{routes_name}")
         json_output = json.loads(output)
         if not json_output["spec"]["host"]:
             return None
@@ -508,65 +278,34 @@ class OpenShiftAPI:
             return None
         return json_output["spec"]["host"]
 
-    def is_pod_ready(self, cycle_count: int = 60) -> bool:
-        """
-        Function checks if pod with specific name is really ready
-        """
-        print("Check if pod is ready.")
-        for count in range(cycle_count):
-            print(".", end="")
-            json_data = self.get_pod_status()
-            if len(json_data["items"]) == 0:
-                time.sleep(3)
-                continue
-            if not self.is_pod_finished(pod_suffix_name=self.pod_name_prefix):
-                time.sleep(3)
-                continue
-            for item in json_data["items"]:
-                pod_name = item["metadata"]["name"]
-                status = item["status"]["phase"]
-                if "deploy" in pod_name:
-                    continue
-                if item["status"]["phase"] == "Running":
-                    print(f"\nPod with name {pod_name} is running {status}.")
-                    output = OpenShiftAPI.run_oc_command(
-                        f"logs {pod_name}", namespace=self.namespace, json_output=False
-                    )
-                    print(output)
-                    # Wait couple seconds for sure
-                    time.sleep(10)
-                    return True
-                time.sleep(3)
-        return False
-
     def get_openshift_args(self, oc_args: List[str]) -> str:
         return " -p ".join(oc_args)
 
     def template_deployed(self, name_in_template: str = "") -> bool:
-        if not self.is_build_pod_finished():
+        if not self.openshift_ops.is_build_pod_finished():
             print("\nBuild pod does not finished in proper time")
-            self.print_get_status()
+            self.openshift_ops.print_get_status()
             return False
-        if not self.is_pod_running(pod_name_prefix=name_in_template):
+        if not self.openshift_ops.is_pod_running(pod_name_prefix=name_in_template):
             print("Pod is not running after time.")
-            self.print_get_status()
+            self.openshift_ops.print_get_status()
             return False
         return True
 
     def command_app_run(self, cmd: str, return_output: bool = True) -> str:
         cmd = f"exec command-app -- bash -c \"{cmd}\""
         print(f"command_app_run: {cmd}")
-        cmd_out = self.run_oc_command(
+        cmd_out = run_oc_command(
             cmd=cmd, ignore_error=True, return_output=return_output, json_output=False
         )
         return cmd_out
 
     def create_deploy_command_app(self, image_name: str = "registry.access.redhat.com/ubi8/ubi") -> bool:
         cmd_file = utils.save_command_yaml(image_name=image_name)
-        self.run_oc_command(f"create -f {cmd_file}")
-        if not self.is_pod_running(pod_name_prefix="command-app"):
+        run_oc_command(f"create -f {cmd_file}")
+        if not self.openshift_ops.is_pod_running(pod_name_prefix="command-app"):
             print("create_deploy_command_app: command-app pod is not running after time.")
-            self.print_get_status()
+            self.openshift_ops.print_get_status()
             return False
         output_cmd = self.command_app_run("echo $((11*11))")
         if "121" not in output_cmd:
@@ -609,7 +348,7 @@ class OpenShiftAPI:
             app_param = utils.download_template(template_name=app)
         oc_cmd = f"new-app {tagged_image}~{app_param} --strategy=source --context-dir={context} --name={service_name}"
         try:
-            output = self.run_oc_command(f"{oc_cmd}", json_output=False)
+            output = run_oc_command(f"{oc_cmd}", json_output=False)
             print(output)
         except subprocess.CalledProcessError as cpe:
             print(cpe.output)
@@ -642,7 +381,7 @@ class OpenShiftAPI:
         oc_cmd = f"new-app -f {local_template} --name={app_name} -p NAMESPACE={self.namespace} {openshift_args}"
         print(f"Deploy template by command: oc {oc_cmd}")
         try:
-            output = self.run_oc_command(f"{oc_cmd}", json_output=False)
+            output = run_oc_command(f"{oc_cmd}", json_output=False)
             print(output)
         except subprocess.CalledProcessError:
             return False
@@ -712,7 +451,7 @@ class OpenShiftAPI:
         oc_cmd = f"new-app {local_template} --name {name_in_template} -p NAMESPACE={self.namespace} {openshift_args}"
         print(f"Deploy template by command: oc {oc_cmd}")
         try:
-            output = self.run_oc_command(f"{oc_cmd}", json_output=False)
+            output = run_oc_command(f"{oc_cmd}", json_output=False)
             print(output)
         except subprocess.CalledProcessError:
             return False
@@ -730,7 +469,7 @@ class OpenShiftAPI:
     ) -> bool:
         if not self.create_deploy_command_app(image_name=image_name):
             return False
-        ip_address = self.get_service_ip(service_name=service_name)
+        ip_address = self.openshift_ops.get_service_ip(service_name=service_name)
         cmd = cmd.replace("<IP>", ip_address)
         for count in range(timeout):
             output = self.command_app_run(cmd=cmd, return_output=True)
@@ -748,7 +487,7 @@ class OpenShiftAPI:
             response_code: int = 200,
             max_tests: int = 20
     ) -> bool:
-        ip_address = self.get_service_ip(service_name=name_in_template)
+        ip_address = self.openshift_ops.get_service_ip(service_name=name_in_template)
         url = f"{protocol}://{ip_address}:{port}/"
         print(f"URL address to get internal response is: {url}")
         if not self.create_deploy_command_app():
